requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
openai==1.55.0
audioop-lts>=0.2.1
wavelink==3.4.1
PyNaCl==1.5.0
Pillow
